[bugfix] fix a unsigned/signed bug.
This bug was introduced at
Commit: e06f807c482cba9cc8dfcce47dbfd5cab7037fb3 [e06f807]
Clean up common/misc.h

+++ src/pc88/pc88.cpp
@@ -130,8 +130,8 @@ bool PC88::Init(Draw* _draw, DiskManager* disk, TapeManager* tape) {
 //  1 tick = 10μs
 //
 int PC88::Proceed(uint32_t ticks, uint32_t clk, uint32_t ecl) {
-  clock = std::max(1U, clk);
-  eclock = std::max(1U, ecl);
+  clock = std::max(1, static_cast<int>(clk));
+  eclock = std::max(1, static_cast<int>(ecl));
   return Scheduler::Proceed(ticks);
 }
